pyproject.toml
[tool.poetry]
name = "test-whisper"
version = "0.1.0"
description = ""
authors = ["Pavel Glavan <pglavan1998@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
whisperx = "3.7.4"
pydub = "0.25.1"
pydantic = "2.12.4"
python-dotenv = "1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
